refactor(login): import the effort rules from the SDK directly
The TUI `thinking-levels` module re-exported the SDK rules only so older
import paths kept working. Nothing outside this branch depends on that
path, so the callers now name the SDK and the module keeps just what is
genuinely TUI-bound: the theme colour token and the short label.

The rules also had no test in the package that owns them — only the TUI
suite reached them through the re-export. Their suites move to
`packages/node-sdk/test/thinking-levels.test.ts`, so a second renderer
cannot break the shared rules without a red test in the SDK.
